Move internal classes to specific namespace
