+++ tests/Unit/Internal/InteractsXmlTraitSpecimen.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PhpCfdi\SatWsDescargaMasiva\Tests\Unit\Internal;
+
+use PhpCfdi\SatWsDescargaMasiva\Internal\InteractsXmlTrait;
+
+class InteractsXmlTraitSpecimen
+{
+    use InteractsXmlTrait;
+}
