--- tests/Unit/Shared/InteractsXmlTraitSpecimen.php
@@ -1,12 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace PhpCfdi\SatWsDescargaMasiva\Tests\Unit\Shared;
-
-use PhpCfdi\SatWsDescargaMasiva\Shared\InteractsXmlTrait;
-
-class InteractsXmlTraitSpecimen
-{
-    use InteractsXmlTrait;
-}
